Fixed Samina system and healthbar(again)

+++ Alice In The House of Insanity/Assets/Scripts/MeterBar.cs
@@ -10,7 +10,7 @@ public class MeterBar : MonoBehaviour
 	public Gradient gradient;
 	public Image fill;
 
-	public void SetMaxValue(int health)
+	public void SetMaxValue(float health)
 	{
 		slider.maxValue = health;
 		slider.value = health;
@@ -18,7 +18,7 @@ public class MeterBar : MonoBehaviour
 		fill.color = gradient.Evaluate(1f);
 	}
 
-    public void SetValue(int health)
+    public void SetValue(float health)
 	{
 		slider.value = health;
